requirements.txt
pyserial==3.4
Wave==0.0.2
